refactor: API通信部分をクラスに切り出しDIのモジュールに登録しました
